Long에서 int

+++ src/main/java/com/izo/itaportal/model/Admin.java
@@ -5,12 +5,12 @@ import lombok.Data;
 @Data
 public class Admin {
 
-    private Long idAdmin;
+    private int idAdmin;
     private String name;
     private String birth;
     private String tel;
     private String address;
-    private Long idUser;
+    private int idUser;
     private String gender;
-    private Long idUser2;
+    private int idUser2;
 }
